refactor: Step 1 구현체에 따른 파일 분리, 이름 수정
- Progress, StudentDictionary 생성 및 해당 구현체 이동
- util.swift -> Util로 이름 수정
